Usernames are now case-insensitive
This fixes a UX issue, where if you accidentally capitalized a letter in the username when logging in, it would instead create a new account. 

The behavior was confusing, since to the user it looks as if their characters were deleted or progress was not saved.

In order for this to work, duplicate accounts (e.g. username and USERNAME) need to be deleted/renamed. The server will *detect* if any duplicates exist. If any are found, it will direct the server operator to a pruning script, or they can choose to resolve the duplicates manually.